% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltr-clus.R
\name{ltr_clus}
\alias{ltr_clus}
\title{left to right clustering}
\usage{
ltr_clus(Dat, k)
}
\arguments{
\item{Dat}{a data matrix}

\item{k}{number of clusters}
}
\value{
a (left to right) clustering vector
}
\description{
A left to right initial clustering method using the
mean differences and Hotelling's T-squared test
}
\examples{
data(CMAPSS)
clus = ltr_clus(CMAPSS$train$x[1:CMAPSS$train$N[1],],3)

}
\author{
Morteza Amini, \email{morteza.amini@ut.ac.ir}, Afarin Bayat,  \email{aftbayat@gmail.com}
}
