% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ltr_reg_clus.R
\name{ltr_reg_clus}
\alias{ltr_reg_clus}
\title{left to right linear regression clustering}
\usage{
ltr_reg_clus(Dat, k, resp.ind = 1)
}
\arguments{
\item{Dat}{a data matrix}

\item{k}{number of clusters}

\item{resp.ind}{the column indices of the response variables for the linear regression clustering approach. The
default is 1, which means that the first column is the univariate response variable}
}
\value{
a (left to right) clustering vector
}
\description{
A left to right initial linear regression clustering method using the
coefficient differences and Hotelling's T-squared test
}
\author{
Morteza Amini, \email{morteza.amini@ut.ac.ir}, Afarin Bayat,  \email{aftbayat@gmail.com}
}
