% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmixar.R
\name{rmixar}
\alias{rmixar}
\title{Random data generation from the mixture of Gaussian linear (Markov-switching) autoregressive models for hhsmm model}
\usage{
rmixar(j, model, x)
}
\arguments{
\item{j}{a specified state}

\item{model}{a \code{\link{hhsmmspec}} model}

\item{x}{the previous x vector as the covariate of the autoregressive model}
}
\value{
a random matrix of observations from mixture of Gaussian linear
(Markov-switching) autoregressive model
}
\description{
Generates vectors of observations
from mixture of Gaussian linear (Markov-switching) autoregressive
model in a specified state and using the
parameters of a specified model
}
\author{
Morteza Amini, \email{morteza.amini@ut.ac.ir}
}
